messy refactor of colors to improve performance and fix dark mode colors.

+++ app/src/main/java/com/joshmermelstein/loopoverplus/GameBoard.kt
@@ -19,8 +19,8 @@ class GameBoard(private val board: Array<Array<GameCell>>) {
         Array(numRows) { row ->
             Array(numCols) { col ->
                 makeGameCell(
-                    col,
-                    row,
+                    col.toDouble(),
+                    row.toDouble(),
                     numRows,
                     numCols,
                     contents[row * numCols + col],
